replace name servlet and create new sql
