--- src/main/java/controler/MainServlet.java
@@ -1,52 +1,0 @@
-package controler;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet(name = "mainServlet",value = "/main")
-public class MainServlet extends HttpServlet {
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String action = req.getParameter("action");
-        if (action == null) {
-            action = "";
-        }
-        switch (action) {
-            case "block":
-
-                break;
-        }
-    }
-
-
-
-    public class AdminServlet extends HttpServlet {
-
-        @Override
-        protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-            String actionGet = req.getParameter("actionGet");
-            if (actionGet == null) {
-                actionGet = "";
-            }
-            switch (actionGet) {
-                default:
-            }
-        }
-
-        @Override
-        protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-            String actionPost = req.getParameter("actionPost");
-            if (actionPost == null) {
-                actionPost = "";
-            }
-            switch (actionPost) {
-                default:
-            }
-        }
-    }
-}
