+++ src/main/java/controler/SessionServlet.java
@@ -118,7 +118,7 @@ public class SessionServlet extends HttpServlet {
 
                 // Chuyển hướng người dùng đến trang tương ứng với vai trò
                 if (permission.equals("admin")) {
-                    resp.sendRedirect("/home");
+                    resp.sendRedirect("/admin");
                 } else {
                     resp.sendRedirect("/user");
                 }
