+++ src/main/webapp/admin/home.jsp
@@ -96,7 +96,7 @@
                             <td>${user.password}</td>
                             <td>${user.permission}</td>
                             <td>${user.status}</td>
-                            <td><a href="home?action=block&id=${user.id}"><i class="bi bi-ban"></i></a></td>
+                            <td><a href="admin?action=block&id=${user.id}"><i class="bi bi-ban"></i></a></td>
                             <td></td>
                         </tr>
                     </c:forEach>
